azureeventgridsendevent/requirements.txt
azure-eventgrid
azure-identity
